Feat: 계약서 삭제 처리
1. 계약서 삭제요청 전체 조회
2. 계약서 삭제요청 단일 조회
3. 계약서 삭제 처리

+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/aggregate/SalesOppDeleteRequest.java
@@ -17,7 +17,7 @@ public class SalesOppDeleteRequest {
     @Id
     @Column(name = "sales_opp_delete_request_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int salesOppDeleteRequestId;
+    private long salesOppDeleteRequestId;
 
     @Column(name = "request_content")
     private String requestContent;
